Create and Update DB reader and writer for test
