+++ PianoManager/db_for_test.py
@@ -1,0 +1,27 @@
+# 데이터 방식이 어떨지 모르지만 임의로 만든 예시 데이터.
+# 세계적인 피아니스트가 꿈인 진영이의 연습량을 표현해보았다.
+# 평범한 김정원과 비교해보자.
+data = [
+    {
+        "ID": "2019123456",
+        "name": "이진영",
+        "contact": "010-1234-5678",
+        "time_used": "17:01:25",
+        "NFC_UID": None
+    },
+    {
+        "ID": "2019234567",
+        "name": "김정원",
+        "contact": "010-1111-1112",
+        "time_used": "00:30:12",
+        "NFC_UID": "81FCF24F"
+    }
+]
+temp_data = {
+    "ID": "-1",
+    "name": "None",
+    "contact": "010-3453-2432",
+    "time_used": "00:00:00",
+    "NFC_UID": "-1",
+    "is_temporary": False
+}
